% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxCoxChord.R
\name{boxCoxChord}
\alias{boxCoxChord}
\title{Compute the Box-Cox.chord transformation on quantitative community
composition data for any exponent. Usual exponents are larger than or equal
to 0.}
\usage{
boxCoxChord(mat, bc.exp = 0)
}
\arguments{
\item{mat}{Matrix or data.frame of quantitative non-negative community
composition data (frequencies, biomasses, energy measures, etc.)}

\item{bc.exp}{Box-Cox exponent to the data before chord transformation.
Usual exponent values are {1, 0.5, 0.25, 0}, where
bc.exp=1: no transformation;
bc.exp=0.5: square-root transformation;
bc.exp=0.25: fourth-root (or double square-root) transformation;
bc.exp=0: log(y+1) transformation (default value).
Default value: bc.exp=0 (log(y+1) transformation).}
}
\value{
A Box-Cox+chord transformed matrix of the same size as the original data matrix.
}
\description{
Compute the Box-Cox.chord transformation on quantitative community
composition data for any exponent. Usual exponents are larger than or equal
to 0.
}
\details{
This code is a minor rewrite of Pierre Legendre's code provided in its
lecture (Appendix 4)
}
\examples{
boxCoxChord(matrix(runif(100), 10, 10))
}
\references{
\itemize{
\item \url{https://en.wikipedia.org/wiki/Power_transform}
\item \url{https://onlinestatbook.com/2/transformations/box-cox.html}
}
}
