% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chalumeau.R
\name{chalumeau}
\alias{chalumeau}
\title{Reproduce Chalumeau analysis}
\usage{
chalumeau(bc.exp = 1)
}
\arguments{
\item{bc.exp}{Box-Cox exponent (see \code{\link[=boxCoxChord]{boxCoxChord()}}).}
}
\description{
Reproduce Chalumeau analysis
}
