% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climatic_var}
\alias{climatic_var}
\title{Climatic variables}
\format{
A data frame with 12,732 rows and 20 columns
}
\usage{
climatic_var
}
\description{
Climatic variables
}
\keyword{datasets}
