% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deposits}
\alias{deposits}
\title{Conversion table for deposit groups}
\format{
A data frame with 3 and 104 rows
\itemize{
\item \code{code}: species code (3 letters)
\item \code{group}: group name
\item \code{frequency}: frequency in POE data
}
}
\usage{
deposits
}
\description{
Conversion table for deposit groups
}
\references{
\url{https://www.donneesquebec.ca/recherche/dataset/points-observation-ecologique/resource/f21dfb59-cf61-4c59-a051-b16435c58c15},
groups by Pierre Grondin
}
\keyword{datasets}
