% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{domain_bio}
\alias{domain_bio}
\title{Biological domains}
\format{
A sf object with 6 domains and 8 columns.
}
\usage{
domain_bio
}
\description{
Biological domains
}
\references{
\url{https://www.donneesquebec.ca/recherche/fr/dataset/systeme-hierarchique-de-classification-ecologique-du-territoire}
}
\keyword{datasets}
