% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{environ_var}
\alias{environ_var}
\title{Environment variables}
\format{
A data frame with 8,581 rows and 19 columns
}
\usage{
environ_var
}
\description{
Environment variables
}
\keyword{datasets}
