% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plot_location}
\alias{plot_location}
\title{A dataset containing plot locations.}
\format{
A sf object with 12,731 points and 6 fields (columns):
\itemize{
\item ``CLE_ECO`:
\item ``LONGITUDE`:
\item ``LATITUDE`:
\item ``wc20bio30s0`:
\item ``wc20bio30s1`:
\item ``TMA`: Temperature annuelle moyenne
}
}
\usage{
plot_location
}
\description{
A dataset containing plot locations.
}
\keyword{datasets}
