% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poe_env}
\alias{poe_env}
\title{POE Environment variables}
\format{
A sf object with 28398 points and 9 fields:
\itemize{
\item id_poe: POE identifier
\item elevation: elevation
\item slope_type: slope type, one among the following values
\itemize{
\item 0: plat
\item 2: escarpement
\item 3: sommet
\item 4: haut_pente
\item 5: mi_pente
\item 6: replat
\item 7: bas_pente
\item 8: depression_o
\item 9: depression_f
}
\item humus_type: humus type, one among the following values
\itemize{
\item MR
\item SO
\item TO
\item MD
\item MS
\item NA
\item MU
\item AN
}
\item mo_thickness: organic matter thickness
\item deposit_type: deposit type (10)
\item deposit_group: deposit group
\item mean_pcp: mean annual precipitation (averaged over 1990-1999)
\item mean_temp: mean annual temperature (averaged over 1990-1999)
\item potential_vegetation: potential natural vegetation
}
}
\usage{
poe_env
}
\description{
POE Environment variables
}
\references{
\itemize{
\item POE data: \url{https://www.donneesquebec.ca/recherche/dataset/points-observation-ecologique/resource/f21dfb59-cf61-4c59-a051-b16435c58c15}
\item Temperature and precipitation data: McKenney, Daniel W., Michael F. Hutchinson, Pia Papadopol, Kevin Lawrence, John Pedlar, Kathy Campbell, Ewa Milewska, Ron F. Hopkinson, David Price, and Tim Owen. “Customized Spatial Climate Models for North America.” Bulletin of the American Meteorological Society 92, no. 12 (December 2011): 1611–22. https://doi.org/10.1175/2011BAMS3132.1.
}
}
\keyword{datasets}
