% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poe_species}
\alias{poe_species}
\title{POE species variables}
\format{
A data frame with 28396 rows and 357 columns. The first column is
the POE identifier (\code{id_poe}) and the rest of the columns are the species
coverages expressed as a percentage (that can be greater than 100\%), the
column names are the species three-letters code.
}
\usage{
poe_species
}
\description{
POE species variables
}
\keyword{datasets}
