% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poe_type_eco}
\alias{poe_type_eco}
\title{POE}
\format{
A sf object with 28398 points and 43 columns.
}
\usage{
poe_type_eco
}
\description{
Point d'observation \u00e9cologique.
}
\references{
\itemize{
\item \url{https://www.donneesquebec.ca/recherche/dataset/points-observation-ecologique/resource/f21dfb59-cf61-4c59-a051-b16435c58c15}, we used the layer \code{vue_pe_type_eco}.
}
}
\keyword{datasets}
