% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_codes}
\alias{species_codes}
\title{Species codes}
\format{
A data frame with 326rows and 4 columns
\itemize{
\item \code{code}: species code (3 letters)
\item \code{francais}: french name
\item \code{latin}: latin name
\item \code{strate}
}
}
\usage{
species_codes
}
\description{
Species codes
}
\keyword{datasets}
