% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triplot_rda.R
\name{triplot_rda}
\alias{triplot_rda}
\title{Custom rda triplot}
\usage{
triplot_rda(res_rda)
}
\arguments{
\item{res_rda}{a rda object}
}
\description{
Custom rda triplot
}
\references{
Code borrowed and adapted from:
\itemize{
\item \url{https://r.qcbs.ca/workshop10/book-en/redundancy-analysis.html}
}
}
